Refactor app module
